@@ -1,6 +1,6 @@
 module github.com/openstack-k8s-operators/lib-common/modules/database
 
-go 1.18
+go 1.19
 
 require (
 	github.com/openstack-k8s-operators/lib-common/modules/common v0.0.0-20220909175216-e774739df18a
