@@ -19,7 +19,6 @@ package util
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
@@ -78,15 +77,14 @@ func GetTemplatesPath() string {
 	return templatesPath
 }
 
-//
 // GetAllTemplates - get all template files
 //
 // The structur of the folder is, base path, the kind (CRD in lower case),
-// - path - base path of the templates folder
-// - kind - sub folder for the CRDs templates
-// - templateType - TType of the templates. When the templates got rendered and added to a CM
-//   this information is e.g. used for the permissions they get mounted into the pod
-// - version - if there need to be templates for different versions, they can be stored in a version subdir
+//   - path - base path of the templates folder
+//   - kind - sub folder for the CRDs templates
+//   - templateType - TType of the templates. When the templates got rendered and added to a CM
+//     this information is e.g. used for the permissions they get mounted into the pod
+//   - version - if there need to be templates for different versions, they can be stored in a version subdir
 //
 // Sub directories inside the specified directory with the above parameters get ignored.
 func GetAllTemplates(path string, kind string, templateType string, version string) []string {
@@ -123,10 +121,12 @@ func GetAllTemplates(path string, kind string, templateType string, version stri
 // execute it with the specified data
 func ExecuteTemplate(templateFile string, data interface{}) (string, error) {
 
-	b, err := ioutil.ReadFile(templateFile)
+	b, err := os.ReadFile(templateFile)
 	if err != nil {
+
 		return "", err
 	}
+
 	file := string(b)
 
 	renderedTemplate, err := ExecuteTemplateData(file, data)
@@ -181,7 +181,7 @@ func ExecuteTemplateFile(filename string, data interface{}) (string, error) {
 		filepath = path.Join(templates + filename)
 	}
 
-	b, err := ioutil.ReadFile(filepath)
+	b, err := os.ReadFile(filepath)
 	if err != nil {
 		return "", err
 	}
